move all code files to src
