+++ config.json
@@ -1,1 +1,1 @@
-{"AWS_ACCESS_KEY_ID":"","AWS_SECRET_ACCESS_KEY":"","ENDPOINT":"","osdPath":""}
+{"AWS_ACCESS_KEY_ID":"","AWS_SECRET_ACCESS_KEY":"","ENDPOINT":"","OSD_PATH":""}
