+++ package.json
@@ -2,7 +2,7 @@
   "name": "dashboards-desktop",
   "version": "1.0.0",
   "description": "<img src=\"https://opensearch.org/assets/brand/SVG/Logo/opensearch_dashboards_logo_darkmode.svg\" height=\"64px\"/>",
-  "main": "main.js",
+  "main": "src/main.js",
   "scripts": {
     "start": "electron ."
   },
